web_server addon mode init data_path options param

+++ src/emhass/web_server.py
@@ -98,7 +98,7 @@ def action_call(action_name):
     runtimeparams = request.get_json(force=True)
     params = json.dumps(params)
     if runtimeparams is not None and runtimeparams != '{}':
-        app.logger.info("Passed runtime parameters: " + runtimeparams)
+        app.logger.info("Passed runtime parameters: " + str(runtimeparams))
     runtimeparams = json.dumps(runtimeparams)
     ActionStr = " >> Setting input data dict"
     app.logger.info(ActionStr)
@@ -227,6 +227,9 @@ if __name__ == "__main__":
         else:
             app.logger.error("options.json does not exists")
         DATA_PATH = os.getenv("DATA_PATH", default="/app/data/")
+        #if data path specified by options.json
+        if options.get('data_path', None) != None and options.get('data_path', None) != "default":
+            DATA_PATH = options.get('data_path', None);      
     else:
         if use_options:
             OPTIONS_PATH = os.getenv('OPTIONS_PATH', default="/app/options.json")
@@ -241,7 +244,6 @@ if __name__ == "__main__":
             options = None
         CONFIG_PATH = os.getenv("CONFIG_PATH", default="/app/config_emhass.yaml")
         DATA_PATH = os.getenv("DATA_PATH", default="/app/data/")
-
     config_path = Path(CONFIG_PATH)
     data_path = Path(DATA_PATH)
     
@@ -373,8 +375,11 @@ if __name__ == "__main__":
         params = build_params(params, params_secrets, options, 1, app.logger)
     else:
         params = build_params(params, params_secrets, options, args.addon, app.logger)
-    with open(str(data_path / 'params.pkl'), "wb") as fid:
-        pickle.dump((config_path, params), fid)
+    if Path(str(data_path / 'params.pkl')).is_file(): 
+        with open(str(data_path / 'params.pkl'), "wb") as fid:
+            pickle.dump((config_path, params), fid)
+    else: 
+        raise Exception("missing: " + str(data_path))        
 
     # Define logger
     #stream logger
